Split Partitioner I/O off into separate headers
